+++ test-utils/src/lib.rs
@@ -19,8 +19,8 @@ pub use vm_core::chiplets::hasher::{hash_elements, STATE_WIDTH};
 
 pub use assembly::{Library, MaslLibrary};
 pub use processor::{
-    AdviceInputs, AdviceProvider, DefaultHost, ExecutionError, ExecutionOptions, ExecutionTrace,
-    ContextId, Process, ProcessState, StackInputs, VmStateIterator,
+    AdviceInputs, AdviceProvider, ContextId, DefaultHost, ExecutionError, ExecutionOptions,
+    ExecutionTrace, Process, ProcessState, StackInputs, VmStateIterator,
 };
 pub use prover::{prove, MemAdviceProvider, ProvingOptions};
 pub use test_case::test_case;
@@ -175,9 +175,8 @@ impl Test {
         // validate the memory state
         for data in expected_mem.chunks(WORD_SIZE) {
             // Main memory is zeroed by default, use zeros as a fallback when unwrap to make testing easier
-            let mem_state = process
-                .get_mem_value(ContextId::root(), mem_start_addr)
-                .unwrap_or(EMPTY_WORD);
+            let mem_state =
+                process.get_mem_value(ContextId::root(), mem_start_addr).unwrap_or(EMPTY_WORD);
 
             let mem_state = stack_to_ints(&mem_state);
             assert_eq!(
